ai-engine/requirements.txt
bytewax==0.19.0
river==0.21.0
pandas==2.2.0
kafka-python==2.0.2
orjson==3.9.10
clickhouse-driver==0.2.6
